+++ hw-2/kmeans_libs/src/kmeans.c
@@ -30,7 +30,7 @@
  */
 int SquareEuclideanDistance(Point const* point_a,
                             Point const* point_b,
-                            float* out) {
+                            int* out) {
   if ((out == NULL) || (point_a == NULL) || (point_b == NULL)) {
     return FAILURE;
   }
